beginning of re-restructure
