+++ models/index.js
@@ -1,7 +1,7 @@
 const Shop = require('./Shop');
 const Bakedgoods = require('./Bakedgoods');
-const Recipes = require('./Recipes');
-const Ingredients = require('./Ingredients');
+// const Recipes = require('./Recipes');
+// const Ingredients = require('../archive/Models/Ingredients');
 const User = require('./User');
 
 Shop.hasMany(Bakedgoods, {
@@ -12,21 +12,21 @@ Bakedgoods.belongsTo(Shop, {
   foreignKey: 'shop_id',
 });
 
-Bakedgoods.hasOne(Recipes, {
-  foreignKey: 'baked_id',
-});
+// Bakedgoods.hasOne(Recipes, {
+//   foreignKey: 'baked_id',
+// });
 
-Recipes.belongsTo(Bakedgoods, {
-  foreignKey: 'baked_id',
-});
+// Recipes.belongsTo(Bakedgoods, {
+//   foreignKey: 'baked_id',
+// });
 
-Recipes.hasOne(Ingredients, {
-  foreignKey: 'recipe_id',
-});
+// Recipes.hasOne(Ingredients, {
+//   foreignKey: 'recipe_id',
+// });
 
-Ingredients.belongsTo(Recipes, {
-  foreignKey: 'recipe_id',
-});
+// Ingredients.belongsTo(Recipes, {
+//   foreignKey: 'recipe_id',
+// });
 
 
-module.exports = { Shop, Bakedgoods, Recipes, Ingredients, User };
+module.exports = { Shop, Bakedgoods, User };
