--- seeds/shopData.json
@@ -1,14 +1,0 @@
-[
-    {
-        "Name": "Bagels",
-        "Description": ""
-    },
-    {
-        "Name": "Cakes",
-        "Description": ""
-    },
-    {
-        "Name": "Cookies",
-        "Description": ""
-    }
-]
